[Build] rename Manager.cpp/hpp to ECSManager.cpp/hpp
