+++ src/client/ClientGame/ClientGame.hpp
@@ -9,7 +9,7 @@
 
 #include "../../WindowsGuard.hpp"
 
-#include "../../shared/ECS/Manager.hpp"
+#include "../../shared/ECS/ECSManager.hpp"
 #include "../../shared/MessageQueue/MessageQueue.hpp"
 #include "../../shared/Networking/UdpCommunication.hpp"
 #include "../../shared/Utilities/UUID.hpp"
